requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pyyaml>=6.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0